requirements.txt
Django==5.1.7
environs==14.1.1
psycopg2-binary==2.9.10
requests==2.32.3
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
gevent==24.11.1
gunicorn==23.0.0
drf-yasg==1.21.10
djangorestframework==3.16.1
